# resample: only compute slerp when necessary