--- a/packages/functions/src/resample.ts
+++ b/packages/functions/src/resample.ts
@@ -109,9 +109,13 @@ function optimize(sampler: AnimationSampler, path: GLTF.AnimationChannelTargetPa
 
 			if (interpolation === 'LINEAR' && path === 'rotation') {
 				// Prune keyframes colinear with prev/next keyframes.
-				const sample = slerp(tmp as quat, valuePrev as quat, valueNext as quat, t) as number[];
 				const angle = getAngle(valuePrev as quat, value as quat) + getAngle(value as quat, valueNext as quat);
-				keep = !MathUtils.eq(value, sample, tolerance) || angle + Number.EPSILON >= Math.PI;
+				if (angle + Number.EPSILON >= Math.PI) {
+					keep = true;
+				} else {
+					const sample = slerp(tmp as quat, valuePrev as quat, valueNext as quat, t) as number[];
+					keep = !MathUtils.eq(value, sample, tolerance);
+				}
 			} else if (interpolation === 'LINEAR') {
 				// Prune keyframes colinear with prev/next keyframes.
 				const sample = vlerp(tmp, valuePrev, valueNext, t);