Currently when resampling quaternion keyframes, both `angle` and `sample` are computed, but not all of them are used for `keep`, for example, if result of `angle` is larger `Math.PI - Number.EPSILON`, value of `sample` is not needed since `keep` would be always true.

Alternative:
Only compute `angle` when necessary, compute `sample` using slerp first, and compute `angle` if they are close enough.

See also <https://github.com/donmccurdy/glTF-Transform/issues/501>