@donmccurdy 2023-04-16
Thank you @kzhsw! Please see feedback in https://github.com/donmccurdy/glTF-Transform/issues/908#issuecomment-1510475423.