@kzhsw 2023-04-17
![before](https://user-images.githubusercontent.com/120431159/232357899-6996be29-a910-428e-8a7c-70886db768d3.png)
![after](https://user-images.githubusercontent.com/120431159/232354559-64adf412-2d01-44ae-a5f3-5e33e5431f56.png)

| function | before (ms) | after (ms) | diff (ms) |
| :---: | ---: | ---: | ---: |
| on (getAngle) |144.9 | 179.1 |  +34.2|
| rn (slerp)| 264.4 | 196.4  | -68.0|
| total| 409.3 | 375.5 | -33.8|

Benchmarked using the html code in #908.