@donmccurdy 2023-04-17
@kzhsw Sorry to hold this up — When I test the `Arm.glb` file mentioned in the issue, with `console.time` blocks around resample(), total time is essentially the same before and after this change. If I check "total blocking time" at the bottom of your screenshots above, it appears similar before and after as well. I see the functions you mention are consuming less time, but I'm not sure the `resample()` function is faster as a result.

Would it be possible to share a glTF file that benefits from the change, in case I need to refactor this code later? Or have the changes in #910 done what was required here?