@kzhsw 2023-04-17
Yes, that is a minor optimization comparing to #910, after which [GraphNode#get()](https://github.com/donmccurdy/property-graph/blob/935f4b6b89cb7d232e3d6a5a4374fb5ab573962a/src/graph-node.ts#L178) is taking most of the CPU time, feel free to close this.

Here is the screenshot of 20 runs in chromium, using console.time:
before
![before](https://user-images.githubusercontent.com/120431159/232384332-bcaf80ee-d6ec-4e25-a4d6-6e3623b5d72f.png)
after
![after](https://user-images.githubusercontent.com/120431159/232383798-b59315e3-fc00-4228-989c-f092145eb477.png)