@donmccurdy 2023-04-17
Ok, I see - thank you! I'll close this out for now, but remain open to ways of improving performance here.